requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
requests>=2.26.0
Pillow>=8.3.1
imagehash>=4.2.1
python-Levenshtein>=0.12.2